# openapi.yaml
openapi: 3.0.3
info:
  title: Pera Agent Wallet API
  version: 0.1.0
  description: "Multi-user REST + MCP API for the Pera dashboard and for AI
    agents. Auth: passkey (WebAuthn) → owner session `ps_…` (everything); scoped
    agent tokens `pat_…` (scopes read | pay | fund | admin) for AI runtimes; the
    static API_BEARER_TOKEN is the admin credential. All amounts are decimal
    USDC strings; every mutation returns tx hashes and emits a per-user event."
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OnrampBody:
      type: object
      properties:
        amountTry:
          type: string
          pattern: ^\d+(\.\d{1,2})?$
      required:
        - amountTry
      additionalProperties: false
    BankTransferBody:
      type: object
      properties:
        iban:
          type: string
          minLength: 10
          maxLength: 40
        reference:
          type: string
          minLength: 4
          maxLength: 40
        amountTry:
          type: string
          pattern: ^\d+(\.\d{1,2})?$
      required:
        - iban
        - reference
        - amountTry
      additionalProperties: false
    OfframpBody:
      type: object
      properties:
        amountUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
      required:
        - amountUsdc
      additionalProperties: false
...
paths:
...
  /onramp:
    post:
      summary: SEP-6 TRY deposit into the user's treasury + simulated wire
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnrampBody"
      responses:
        "202":
          description: "{ anchorTxId, status, instructions, destination }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400": *a1
  /onramp/{id}:
    get:
      summary: Anchor transaction passthrough
      security: *a2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Anchor transaction
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /offramp:
    post:
      summary: SEP-6 withdraw USDC → TRY (vault → treasury first if needed; sponsor
        pays)
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfframpBody"
      responses:
        "200":
          description: "{ anchorTxId, stellarTxHash, status, amountTry }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400": *a1
...
  /onramp/instructions:
    post:
      summary: "The user's deposit details: anchor IBAN + the reference that routes a
        bank transfer to their treasury (a SEP-6 order, reused until paid into)"
      security: *a2
      responses:
        "200":
          description: "{ iban, bankName, reference, anchorTxId, minTry, maxTry }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /bank/transfer:
    post:
      summary: "Sandbox bank rail (public): pay amountTry to an IBAN with a reference
        — plays the bank for the mock anchor (pnpm bank)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankTransferBody"
      responses:
        "202":
          description: "{ reference, amountTry, status, anchorTxId }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": *a1
        "409": *a1
...